flask_app/requirements.txt
flask==1.1.2
python-dotenv==0.14.0
flask-sqlalchemy==2.4.4
psycopg2==2.8.5
flask-migrate==2.5.3
flake8==3.8.3
SQLAlchemy-Utils==0.36.8
flask-wtf==0.14.3
email-validator==1.1.1



